if -f

+++ install/install.sh
@@ -650,9 +650,14 @@ iface=$(ifconfig | grep -B3 "inet .*$ip" | grep '^[a-zA-Z0-9].*' | awk '{print $
 
 if [[ -n "$iface" ]]; then
     prettyPrint "IPv4: $ip and interface: $iface"
-    if ! grep -E '^interface:\S+' "$HOME/.iftop.conf"; then
-        cp "$ZPWR_INSTALLER_DIR/install/.iftop.conf" "$HOME"
-        echo "interface:$iface" >> "$HOME/.iftop.conf"
+    if [[ -f "$HOME/.iftop.conf" ]]; then
+        if ! grep -E '^interface:\S+' "$HOME/.iftop.conf"; then
+            cp "$ZPWR_INSTALLER_DIR/install/.iftop.conf" "$HOME"
+            echo "interface:$iface" >> "$HOME/.iftop.conf"
+        fi
+    else
+            cp "$ZPWR_INSTALLER_DIR/install/.iftop.conf" "$HOME"
+            echo "interface:$iface" >> "$HOME/.iftop.conf"
     fi
 fi
